backend/requirements.txt
# Python依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
alembic==1.13.0
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI依赖
openai==1.6.1
langchain==0.1.5
langchain-openai==0.0.6
chromadb==0.4.22

# 爬虫依赖
scrapy==2.11.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.4
pandas==2.1.4
numpy==1.26.3

# 消息队列
aiokafka==0.10.0

# 工具依赖
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
